Remove VoterBeta and fix wait after loop end
